requirements.txt
bcrypt==4.2.0
email_validator==2.2.0
fastapi==0.112.2
httpx==0.27.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.2
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.32
uvicorn==0.30.6
